ai-chat-interface/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.4
PyJWT==2.8.0
bcrypt==4.1.2
gevent==23.7.0
eventlet==0.33.3